+++ src/test/java/br/fakecompany/projectX/DeniniTest.java
@@ -62,33 +62,4 @@ public class DeniniTest
     public void testFail(){
         assertTrue(true);
     }
-
-    /*
-     * @Test
-     * public void testFactorialGabrielLuiz(){        
-     *  try{
-     *     myUtils.factorial(-1);
-     *  }catch(Exception ex){
-     *      String resultado = ex.getMessage();
-     *      String esperado  = "Número deve ser maior ou igual a zero";
-     *      assertEquals(esperado,resultado);
-     *  }
-     */
-    @Test(expected = IllegalArgumentException.class)
-    public void testFactorialGabrielLuiz(){        
-            myUtils.factorial(-1);
-    }
-
-    @Test
-    public void testConverterNumerosRomanosRecebeStringVaziaGabrielLuiz(){
-        int resultado = myUtils.converterNumerosRomanos("");
-        int esperado  = -1;
-        assertEquals(esperado, resultado);
-    }
-    @Test
-    public void testConverterNumerosRomanosRecebeNullGabrielLuiz(){
-        int resultado = myUtils.converterNumerosRomanos(null);
-        int esperado  = -1;
-        assertEquals(esperado, resultado);
-    }
 }
